+++ tidy3d/plugins/adjoint/components/data/monitor_data.py
@@ -32,7 +32,7 @@ class JaxMonitorData(MonitorData, JaxObject, ABC):
     def from_monitor_data(cls, mnt_data: MonitorData) -> JaxMonitorData:
         """Construct a :class:`.JaxMonitorData` instance from a :class:`.MonitorData`."""
         self_dict = mnt_data.dict(exclude={"type"}).copy()
-        for field_name in cls.get_jax_field_names():
+        for field_name in cls.get_jax_field_names_all():
             data_array = self_dict[field_name]
             if data_array is not None:
                 coords = {
@@ -287,7 +287,6 @@ class JaxFieldData(JaxMonitorData, FieldData):
             # Offset coordinates by source center since local coords are assumed in CustomCurrentSource
 
             for freq0 in tuple(self.field_components.values())[0].coords["f"]:
-
                 src_field_components = {}
                 for name, field_component in self.field_components.items():
                     field_component = field_component.sel(f=freq0)